Add DropDownMenu method to initialize the menu items
When creating a new DropDownMenu, you has to pass an array with
info about the dropdown menu items. The items of this array are
object with a label and a link field. A null field, means that the
item is a simple label. Otherwise, an <a> element is created,
pointing to the 'link' field.

let menuData = [
    {"label": "Anchor 1", "link":"#"},
    {"label": "Anchor 2", "link":"#"},
    {"label": "Simple Label", "link":null}
];