Add methods and event listeners to hide/show the menu
Methods to toggle the visibility of a given element have been added
to the DropDownMenu class.
- #toggleVisibility method simply toggles a 'visible' class to
  a given HTML element, which simply sets the CSS property 'visibility'
  to 'visible'.
- #toggleForcedVisibility method simply toggles a 'visible-forced'
  class to a given HTML element, which imposes the CSS property 'visibility' to 'visible !important'.

There are two possible ways to hide/show the dropdown menu:

a) by mouse clicking on the menu button, which toggles the (forced)
   visibility of the main content.

b) To show the menu content: by hovering into the main button, which
   sets the (non-forced) visibility of such content.

   To hide the menu content: by hovering out of the whole menu (button
   + content), which removes the (non-forced) visibility of such
   content. Note that if the .visible-forced class is still applied
   from (a), the content is forced to remain visible.

   This possibility (b) is disabled by default, but can be enabled
   by creating the dropdown menu with the enableHover parameter set
   to true.

Clicking on the (visible) dropdown menu content always removes both
the 'visible' and 'visible-forced' classes, so that the content
becomes hidden.

Note that for now the visibility property is used to hide/show the
dropdown menu content. It should be better to change the display
property instead, so that the corresponding event listeners are disabled, too, when the content is hidden. (todo)