Adjust the position of the drop-down menu content with respect to it
To let the dropdown menu content (.drop-down-content element) be
positioned above the page content, and not occupy empty space, we
need an absolute position for it, with respect to the the whole
drop-down menu (.drop-down-menu element), which in turn needs a
relative position.

To include a gap between the button and the content, you can't
simply impose a margin-top for the content .drop-down-content,
since it is absolute positioned.

Then, .drop-down-content has been wrapped into a content container .drop-down-content-cnt, which is the one that actually is absolute
positioned with respect to the .drop-down-menu element. Then, the
gap can be imposed by simply setting a margin-top on the
.drop-down-content.

Note that changing the top property of an absolute positioned
element would break the hovering behaviour of the dropdown menu,
since moving the mouse from the button through the gap created this
way would trigger a mouseleave event before reaching the actual
content (hence the need of the .drop-down-content-cnt element).

To conclude, the display propery of the .drop-down-content is also
set to switch between none and block (as well as its visibility,
between hidden and visible), so that even event listeners are
disabled when the content is hidden.