Add frame div creation in the ImageCarousel div
A frame div (.image-carousel-frame) is created. This is now the
direct child of the .image-carousel div, while the previously
created .image-carousel-slides is its child now.

To fit the current shown image (as of now, the first one) in the
frame, the width of each img has been set to 100%. So, you just
need to specify the width of the .image-carousel div (the width of
the .image-carousel-frame is implicitly 100%).

As the .image-carousel-frame must only show one image, the property
'overflow: hidden' has been set.