Add method to show a specific slide in the image carousel
Add #showSlide(idx) method to the ImageCarousel class to shown in
the frame the slide/image at index idx.

idx is assumed to be a valid index (between 0 and N-1. where N is
the total number of images).

To do the actual switch, the .image-carousel-slides element
containing all the slides (side by side, horizontally) has been set
to position relative, so that you can change the 'left' property
to move this div within the frame.

As all the slides/images have the same width (100% of the frame),
to show the image at index idx, you have to set the 'left' property
to `-${idx*100}%` (eg, for the image at index 4, you need to set
"left:'-400%'").