Add variable to keep track of image currently shown in image carousel
A private variable #currentImg has been included to keep track of
the index of the image currenlty shown on the image carousel.

The total number of images has been stored in another private
variable #numOfImgs, too.

Variable #currentImg needs to be a valid index (a number between 0
and #numOfImgs). If this is not the case, it has to be rescaled to
a valid index. Private method #getValidIdx(idx) has been written
to do that for a generic idx, to make it a valid one.

Now, the #showSlide method also calls #getValidIdx(idx) to get a
valid idx, and sets #currentImg to idx, too.