Add navigation buttons to the image carousel
To navigate (jump) to each image, #numOfImgs buttons are added,
each of which allows showing a specific image.

These buttons are cointained in the new .image-carousel-navigation
div. Note that this div has been absolute positioned, and its
z-index has been set to 1, to make it prevail over the previous and
next buttons.