Make seamless transition from last to first image in the image carousel
Suppose we have N=this.#numOfImages images, for i=1,2,...N. These
are appended, in order, to the .image-carousel-slides div:

[1],[2],[3],...,[N]

We would like to make the cycling through it appear seamless, as
if it were:

[1],[2],[3],...,[N],[1],[2],[3],...,[N],[1],[2],[3],...,[N],...

To make a seamless transition N -> 1 when pushing the next button
or on automatic slide cycling, we need to append the first image
at the end of .image-carousel-slides div (but don't count it in
the this.#numOfImages):

[1],[2],[3],...,[N],[1']

We consider the following private properties in the ImageCarousel
class:

   - #idxOfFirstImg:        index of [1] (in .image-carousel-slides
                            div), ie, 0

   - #idxOfLastImg:         index of [N], ie, this.#numOfImages-1

   - #idxOfAppendedFirstImg:index of [1'], ie, this.#numOfImages

   - #currentImg: the index of the current image shown, ie {0,1,2,
                  ...,N-1}, corresponding to {[1],[2],[3],...,[N]}.

   - #idxForLeft: the index in .image-carousel-slides div of the
                  currently shown image, ie {#idxOfFirstImg, ...,
                  #idxOfLastImg, #idxOfAppendedFirstImg},
                  corresponding to {[1],[2],[3],...,[N],[1']}.

                  Recall that to show the i-th image in
                  .image-carousel-slides div, you need to set the
                  'left:-i%'' property on .image-carousel-slides div.

The rest of the code has been modified to use these instead of
hardcoded values (eg, using this.#idxOfFirstImg instead of 0).

Then, to do the slide transition N -> 1, first do [N] -> [1']
(setting #currentImg=0 and #idxForLeft=#idxOfAppendedFirstImg, and
applying the transition duration defined in the .css file). Once
this is completed, a 'transitionend' event is fired; then,
temporarily suspend the transition CSS property to make a [1'] ->
[1] transition instantaneously (setting #idxForLeft=#idxOfFirstImg).

Note that an this.#idxOfFirstImg has been explicitly introduced
because the first image might not be on location 0. In fact, to
have a smooth 1->N transition, we need to repeat the above
procedure, prepending image [N] to the images in
.image-carousel-slides div. This is not implemented, yet.