Make seamless transition from first to last image in the image carousel
Referring to the previous commit, we now want to make a seamless
transition N -> 1 when pushing the previous button.

Similar considerations can be made like those for the 1->N
transition.

The main difference is that now we need to prepend the last image
at the beginning of .image-carousel-slides div, so that we get:

[N'],[1],[2],[3],...,[N],[1']

We introduce the private variable #idxOfPrependedLastImg as the
index of [N'] (in .image-carousel-slides div), ie 0.

Variables #idxOfFirstImg, #idxOfLastImg, and #idxOfAppendedFirstImg
are opportunely shifted.

Note that while #currentImg possible values remain the same as
before (ie, {0,1,2,...,N-1}, corresponding to {[1],[2],[3],...,
[N]}), now #idxForLeft can assume one more value (ie,
{#idxOfPrependedLastImg, #idxOfFirstImg, ..., #idxOfLastImg,
 #idxOfAppendedFirstImg}, corresponding to {[N'],[1],[2],[3],...,
 [N],[1']}).