Refactor DropDownMenu.css + change button icon when forced visibility
The code has been rearranged, by ordering the css properties
alphabetically. The selectors have been ordered, too, based on
the element they refer to.

Some modifications have been made to the css variables, to make
it easier to customize the drop down menu.

When the visibility of the content is forced, the icon of the
button is changed (eg, to a X symbol). This is done via JS.

The document has been reformatted using Visual Studio Code, too.