Add index.css with global styles for the webpage
The style definition of the website has been removed from index.js
and moved to index.css. It has also been improved.

In index.css you could also add custom styles for the DropDownMenu
and ImageCarousel object instances. Eg, the following rules
override the default rules of DropDownMenu.css for the first
instance of .drop-down-menu, to make it use red tones:

.drop-down-menu:first-of-type {
    --color-light: peachpuff;
    --color-dark: crimson;
}

Consider adding a css class to the DropDownMenu and ImageCarousel
object instances, to better specify them with a higher specificity,
when overriding their default style.