Fix ImageSlider not adapting to the container in some conditions
The ImageSlider component is set to adapt to the containing
element dimension.

There was a problem by which the ImageSlider component expanded
its coinaining element (<main>, in this case) when (apparently)
its aspect-ratio was 'wider' thatn that of the images of the div.

The problem has been fixed on the external code, actually, and not
on the component (altough some 'width: 100%'' rules have been
added in there for better clarity).

The containing item was a flex element with 'flex: 1'. Since no
min-height was specified, this property defaulted to auto, which
might cause some problems.

See: https://stackoverflow.com/questions/72410543/how-to-use-object-fit-with-flex-grow-flex-shrink

The problem has been fixed by setting 'min-height: 0' on <main>.