Add feature to show more slides in the frame of image carousel
You can now show more slides on the frame of the image carousel.
A specific (percentage) width can be specified, using the
.setImgWidthPercentage(width) method. This is the width of each
slide, with respect to the frame size.

Such width is a integer representing a percentage, which must
satisfy the following constraint:

	max(100/#numOfImgs, 20) <= width <= 100.

If this is not the case, this width is clamped.

Then, in the frame, the current image is shown in the center, and
(100-width)/(2*width) images are shown on each side, too.

Eg, if the width is set to 100 (ie, 100%), on the sides, no image
is shown (indeed, (100-100)/(2*100) = 0). This is the case that
has been considered so far.

Or, if the width is set to 40 (ie, 40%), on the sides, 3/4 of the
next/previous image is shown on the corresponding side (indeed,
(100-40)/(2*40) = 0.75 = 3/4).

Note that right now this parameter is hard-coded in the constuctor
(see this.setImgWidthPercentage(70) on the constructor). This is
going to be obtained from the function input in a future version
(todo).