Improve interface for ImageCarousel class constructor
For the image carousel, you can set the parameters regarding the
functionalities from the interface of the constructor:

	constructor(parentDiv,imagesPaths,options = {}){...}

In particular, now 'options' is an optional object that defines such
properties. The following instucttion is called in the constructor:

	this.#options = Object.assign(defaultOptions,options),

where

	const defaultOptions = {
	  slideTimeoutInMs: 5000,
	  autoCycling: true,
	  imgWidthPercentage: 100,
	  minImgWidthPercentage: 20,
	}

is an object local to the ImageCarousel module that defines the
default values. If 'options' is not specified, it assumes the values
of 'defaultOptions'. Similarly, if any of the field from
'defaultOptions' is not specified in 'options', it assumes the
corresponding default value from 'defaultOptions'.