Improve interface for DropDownMenu class constructor
For the drop-down menu, you can set the parameters regarding the
functionalities from the interface of the constructor:

	constructor(parentDiv, menuData,  options={}){...}

In particular, now 'options' is an optional object that defines such
properties. The following instucttion is called in the constructor:

	this.#options = Object.assign(defaultOptions,options),

where

	const defaultOptions = {
	    enableHover: true,
	    buttonFaIcon: {prefix: 'solid', icon: 'bars'},
	    buttonFaIconForced: {prefix: 'solid', icon: 'xmark'},
	    buttonLabel: '',
	    align: 'center'
	}

is an object local to the DropDownMenu module that defines the
default values. If 'options' is not specified, it assumes the values
of 'defaultOptions'. Similarly, if any of the field from
'defaultOptions' is not specified in 'options', it assumes the
corresponding default value from 'defaultOptions'.