Add module creditFooter.js to add a credit footer to the web page
The default exported funtion setCreditFooter(data) looks for a
<footer> element in the page. If it doesn't find one, it creates
one.

Then, it creates a .credit-footer div with the credit footer inside.
This div is then appended to the <footer>, and returned as well.

This function should be created after the structure of the webpage
has been created, in the end.

'data' is an optional parameter with whom you can specify a
different displayed name and github name, which will be used in
the credit footer. Since the function will be primarly be used for
my personal projects, the default data are my data.

The style of this credit footer is also applied by this function as
an inline style, so no further style is required.

Note: the font family are imported as described here:
https://stackoverflow.com/questions/5586845/how-to-import-font-file-using-javascript

It should be investigated whether using the approach described in
the following page is more viable (since it does not require a direct
font link):
https://github.com/typekit/webfontloader